Check Tuleap Groups directly on Tuleap server (#21)
* Check Tuleap Groups directly on Tuleap server

This is a part of [story #14018](https://tuleap.net/plugins/tracker/?aid=14018) have a central management of users and groups using oauth

From now on, groups are not inferred from token anymore. In order to
test this patch you can:

* Deploy this plugin
* On Tuleap: As a project admin create a group your are not member of in your project
* On Jenkins: add this group in the authorization Matrix: it should
  behave correctly.

You can also try to invalidate tokens on your Tuleap to check that they
are effectively refreshed.