Stop refreshing access tokens (#54)
This change is part of [request #14018](https://tuleap.net/plugins/tracker/?aid=14018)

You shouldn't be prompted for approval for every connection anymore as
we do not require the offline_access anymore. This isn't the correct way
to deal with token expiration on the client side and as long as we don't
have a proper way of handling this (being able to pass the user through
complete reauthentication) we prefer to avoid any black magic to deal
with it.